kronecker tests separated out into own file
