+++ src/probnum/random_variables/_random_variable.py
@@ -10,7 +10,7 @@ from typing import Any, Callable, Dict, Generic, Optional, Tuple, TypeVar, Union
 import numpy as np
 
 from probnum import utils as _utils
-from probnum.types import (
+from probnum.type import (
     RandomStateType,
     ShapeType,
     # Argument Types
